(maint) Maintain naming conventions for puppet msi
In the Puppet 3.x series, we ship two msi's, one with the `-x64` suffix,
and one without. In the Puppet 4.x series, this changes to include the
arch for both packages. Unfortunately, this convention wasn't maintained
in previous changes meant to support to move to including architecutre
in both builds. This commit adds back the naming convention for builds
in the 3.x series by only appending the architecture if the build is
64bit.