Include short sha in puppet-agent MSI filename
Previously, MSIs were generated using `git describe` with the short sha
removed, e.g. `3.7.4-1303`. This version identifier is not passed
through the AIO pipeline, and is not necessarily unique enough to
identify the puppet-agent package to be installed.

This commit changes the xml template for creating jenkins jobs so that
the resulting puppet-agent MSI contains the short sha, e.g.
`3.7.4-1303-ge96f55b`. This  will simplify the process for installing
MSIs during AIO acceptance testing.