(RE-4045) Add support for alternate repositories
Previously the packaging repo was limited to shipping to main/production
or devel, based on the versioning scheme of the project being shipped.
This commit extends that to allow the rpeo to be overriden for apt or
yum repos using a param in the yaml file. apt_repo_name and
yum_repo_name are both available to override the defaults.