(RE-4139) Update ship taks to use repo_command if set
If the user has used the new Pkg::Config.apt_repo_command or
Pkg::Config.yum_repo_command, we want to prefer those commands over the
legacy defaults. This commit adds logic to check if the new param is
set, and if so to invoke it, and otherwise to fall back to the old
behavior. The freight rake task is renamed to update_apt_repo, and the
uber_ship is updated to use it instead of freight.