(RE-4221) Allow building and shipping of only tar/gem artifacts
With the shift to delivering our software via the puppet-agent package,
we want to discontinue shipping conventional packages. However, we do
want to continue shipping the tar and gem artifacts.

This commit lets us maintain a list of LTS build targets for community
members who are using this automation to build rpm or deb packages while
not building rpm/deb packages when we run `rake pl:jenkins:uber_build`.
It also lets us easily not ship any deb/rpm packages. The important this
is that `pl:deb` and `pl:mock` still work for building deb/rpm packages
if that is wanted. If the setting needs to be overridden, like for a
jenkins uber_build, we can do that by setting `RPM=TRUE` and/or
`DEB=TRUE` as environment variables when that rake task is invoked.