(RE-4261) Add build_tar config parameter
For vanagon projects, tarballs will not be produced. This breaks the
uber_ship, which expects a tar to be there. This commit addresses this
by adding a build_tar parameter which defaults to true, which can be
used to toggle the tar ship/build. The TAR environment variable can also
be used to set this on the command line.