(RE-3293) Treat SNAPSHOT builds like RCs for versioning
Previously, ezbake builds would have a SNAPSHOT build that was treated
like any other. This had the consequence of making the 'final' release
for a given version series as less than the SNAPSHOT builds, which means
it would be unpromotable (also not immediately upgradable from the
snapshot version). This commit updates the versioning in packaging to
treat SNAPSHOT builds as if they were RC builds. This means that
SNAPSHOT versions will have a '0.1' prepended to them (0.7.0.SNAPSHOT...
becomes 0.7.0-0.1SNAPSHOT...), which makes the final version higher than
the snapshot version for packaging purposes.