(maint) Allow mock and cow list to be empty strings
Previously, the automation allowed empty strings to used in the `cows`
and `final_mocks` definition. In changes that I made recently, I took
away that functionality. This only impacts repo creating, which fails if
there are no packages of that kind to create a repo for.

This commit makes it so that we will not attempt to create repos in any
of the following situations:

`cows:`
`cows: ''`
`#cows: base-lucid-amd64.cow`

and will attempt to create repos only in the folowing situations:

`cows: base-lucid-amd64.cow`