(RE-4292) Add uber_ship binary to packaging
This commit adds an uber_ship binary to packaging. This file serves two
purposes. First, it allows us to ship ezbake projects in a more
automated fashion than was previously possible. Running `./bin/uber_ship
puppetserver 2.0.0` would ship puppetserver 2.0.0 to our public repos,
without any of the manual steps that are currently required.
The second purpose is to reship a given package. For puppetdb, we will
be shipping the 2.3.x to two repos, products/main and PC1. This was not
possible with our previous tooling. With this commit, if a third
optional argument is included, the packages will be rearranged under the
target repo given, and shipped there. For puppetdb, that means that
2.3.3, which was targeted at main/products, can be rearranged under the
PC1 repo structure and shipped to our public repositories in an
automated fashion, without manual steps required.