(RE-4595) Allow RPMS for SLES and eos to be signed
Currently in packaging you can't ship SLES or eos builds because you
can't sign them, because they were never run through the packaging repo
tasks before. With the advent of vanagon, this is now a requirement, so
here we go.

Sometime in the future, we should consolidate pe_sign.rake and sign.rake
into something that is smarter, but for now this should get us
unblocked.