(RE-4681) Generalize nightly repos tasks
Previously the nightly_repos rake tasks were very specific to a single
purpose, signing and shipping nightly repos. We now have another need
for very similar functionality, so this commit generalizes these tasks
refer to signed repos instead of nightly repos, while allowing the
directory to be named as it was before.