(RE-4681) Enable shipping of builds using the version
For some builds it will be preferable to have autosigned builds shipped
to directories according to their version, and not their ref. This makes
them easier to identify based on an already installed package and will
enable internal ships of puppet-agent to proceed. This commit adds a
new parameter to the deploy_signed_repo task. If versioning is set to
ref, the current behavior is maintained. If it is set to version, then
the repo_configs are updated and the repos are shipped to a directory
named after the dotted version. The deploy_nightly_repos shim is also
updated here to pass ref to maintain current behavior.