(RE-4155) Ensure directories are created with 775 perms
Prior to this commit, directories were being created on
builds.puppetlabs.lan with 755 permissions. This is fine except when
we need to add additional directories to them. For instance, when we do
a release, we create a 'shipped' directory under the main sha/tag
direcotry. This would cause ships to break because it would be unable to
create the 'shipped' directory. This commit updates the ship process to
always create directories with 775 permissions.