(maint) Refactor deploy_signed_repos into two tasks
deploy_signed_repos is a fairly large task that does too much stuff for
a single task. This commit breaks part of the responsibilities out to
lighten it up a little. The ship portion of the deploy is broken out of
the previously named deploy_signed_repos. Accordingly, that task is
renamed to prepare_signed_repos.