(RE-4842) Add archival task to nightly_repos.rake
This commit adds a task to tar up a repo for shipping as part of repo
deployment. This task assumes that prepare_signed_repos has been run
previously for the build. This will be used to deploy tarballs of
signed repos for easy installation of puppet-agent. The archives of
repos could be on a per platform or a per version basis for a given
package. The initial use case is per platform/release archives.