(RE-2542) shipping staging automation for osx
This commit updates the automation to allow shipping of dmgs for vanagon
projects ( vanagon_project set to true ) and when build_dmg is not set.
It also updates the source directory for the rsync to be pkg/apple/
instead of pkg/apple/*.dmg.  This adds support for apple projects that
have a repo defined, such as PC1.