(maint) Do no load variables prematurely
Prior to this commit, the variables used in this file were being loaded
prematurely. As soon as rake was invoked, they would be populated.
However, the correct value was not guarenteed to be there initially. If
we load data from the build-data repo, for instance, it is not
available, because the file is not reloaded and the variables are not
reassigned. This commit ensures that if we load data at a later point in
the process, we will correctly populate these variables.