(maint) Add ability to base64 incode username and password
This is needed to effectively link two tickets together. Since the jira
gem doesn't have support for this functionality yet, we need to flex the
wonderful JIRA API to do it. We don't want to pass in plain text
passwords, so instead, we use basic auth with base64 encoded
information