(maint) Add way more complexity to issue creation
We need to be able to have a lot more complexity with the issues that we
create. This commits adds that in. We can now specify story points,
different issue types (beyond just Task and Sub-task), and more. This
will make it a lot easier to create a complex set of tickets that don't
need any manual updates after they are created with the script.

This commit also updates tickets.rake to comply with these new changes
to issue creation.