(RE-5524) Add Platform module to Pkg and Pkg::Config
In order to reason about platforms and paths in subsequent tasks, the
packaging repo needs to be aware of them in the first place. This commit
takes the hash from enterprise-dist's Rakefile, and applies some massage
to the formatting and content. A few utilities are also added, to get at
expected paths for repos and artifacts for various platforms.