(RE-5524) Add whitelisting ability to repo deployment
This work leverages the previous commit to add platform methods and adds
two new parameters to allow some filtering of which builds are shipped
to the nightlies repo. Here we add a new parameter to the
deploy_signed_repos task, which allows it to operate in a foss_only mode
where only targets listed in the foss platforms array will be shipped.
The two new parameters, foss_platforms and pe_platforms are arrays of
platforms as defined in Pkg::PLATFORM_INFO.