(RE-5573) Update tar creation to set group and owner to root
Previously the tar command used to pack up repos did not specify owner
or group, which meant that it was set to the user running the task,
which is generally jenkins. This user/group won't always exist on the
system that is fetching the tarballs, so this commit updates the tar
invocation to also pass --owner=0 --group=0 to ensure that it is set to
root or root's equivalent when packing.