(RE-5577) Fix osx signing paths
Currently, when signing osx pkgs, when the automation
retrieves the DMGs it's dropping them in the PC1 directory,
meaning when we ship we ship signed packages in the wrong
location, and unsigned packages in the correct location.

This updates the retrieval to put the packages in the correct
location when they get retrieved.