(RE-1198) Use :apt_signing_server if defined
apt_signing_server is a new value that can be added to build_defaults
or build_data to target a new Apt repo signing server, hopefully
divorced of the public shipping infrastructure.

Additionally, the old "run a Rakefile over ssh" code path has been
deprecated and replaced with a warning that updating remote apt
repos is never going to happen.