(maint) Add debug flag to Pkg::Util::Execution.ex
It is currently fairly difficult to get details of why a command has
failed or what it returned by default in the packaging repo. This commit
aims to make that a little easier but adding a parameter to the ex
method to expose what command was executed and what it returned.