(CPR-243) Add ability to sign for Solaris 11
This will allow us to sign Solaris 11 packages from rake tasks in
the packaging repo. Signing for Sol11 is a complicated process, so we
rsync the files up to the signing server, create a new repo there, sign
the repo, pull the packaging files back out, and rsync them back to our
build box. This adds the Pkg::IPS library to be able to sign packages,
and updates the rake task to use the new functionality.