(maint) Add release name to subtask summaries and projected date to descriptions
Whenever a release ticket is created, the subtask assignees receive mystery meat
emails that tell us a release is happening, but not WHICH release, nor when it's
supposed to go down. Also, if there are multiple releases in flight at once (and
in my team there always are), the identical ticket summaries make it really hard
to distinguish between browser tabs.

This commit should hopefully improve that, by:

* Adding a string like " (puppet 4.3.0)" to the end of every release subtask
  summary, so we can distinguish between tickets in emails and tabs.
* Adding the originally projected release date (with caveat) to the beginning of
  every release subtask description, so at least the _initial_ email can tell us
  what the rest of the release crew is thinking.