(RE-6007) Simplify Jira queries
Generates per-component "fixed in" and "introduced in" queries, and adds them
to the ticket for preparing release notes.

Simplifies query logic for tickets needing to be closed at the end of
the release, and the tickets needing to be made public.

Previously, we were incorrectly querying for version "PCP 1.0.2", which
won't work, because PCP has multiple component versions, e.g. pxp-agent
1.0.2, cpp-pcp-client 1.0.1. This commit changes the query to use
"pxp-agent 1.0.2". This means cpp-pcp-client is excluded, but it was
excluded before, so no real change. We can revisit that at a later date.