(RE-5619) Allow shipping to PE feature repos
For PE, we want to allow the promotion and use of arbitrary
test/dev packages into feature branches of PE, that don't pollute
the mainline PE branches and repos.

This adds logic to enable shipping to the feature repos for a PE branch,
and during a normal ship/promotion, links the shipped package into the
feature repos (so that mainline packages are immediately available to the
feature branches).