(RE-6007) Use get_auth_vars to determine the default site
Previously, we duplicated the logic for defaulting the site, because we
needed the site to construct the appropriate queries.

This commit moves the call to `get_auth_vars` earlier so that the
existing logic for defaulting the site is used. It does change the
precedence order, as variables in the rake task have the potential to
overwrite `site` or `user`, but that doesn't happen in practice.

Also removed the misleading comment that calling `get_auth_vars`
authenticates against Jira. That doesn't happen until we call
`validate_agent_release_ticket_vars` which calls `jira.project
vars[:project]`, at which point we login to Jira to resolve the project
with that name.