(RE-6007) Add cpp-pcp-client component to queries
Previously, we were only tracking the pxp-agent component of PCP in the
various Jira queries, but not cpp-pcp-client, which is also a PCP based
component in puppet-agent.

This commit adds a CPP_PCP_CLIENT_RELEASE variable and updates the
various queries to include that component when reporting on tickets in
the release.

Note since cpp-pcp-client is not user facing, we don't provide
component-level "Fixed in" and "Introduced in" queries for
cpp-pcp-client.