(RE-5813) Use PA as the parent project for agent platform tickets
Since the PA project exists we should always use that for platform
additions. We also always want puppet_agent_configuration to be part of
the PA project, and c_toolchain and build tools should be either in PA
or RE depending on whether or not the platform is PE only.