(RE-6070) Ship RPMS to YUM_STAGING_SERVER
We used to ship our RPM repos directly to our perimeter nodes,
which in the new world is sort of not getting the job done anymore.
We should be shipping them to a staging server instead. This commit
adds the pl:remote:deploy_yum_repo method, and updates yum tasks
to use the new 'yum_staging_server' config value.