(MAINT) Refactoring Pkg::Util::Net: move print_url_info
Step 6: Move Pkg::Util::Net.print_url_info out of Net and into
Jenkins. It's not stirctly a networking related method, so it's
being moved to a farm upstate where it can play with other generic
Jenkins related methods until I write an Output or Logging class.

Additionally, the test for Pkg::Util::Net.print_url_info has
been rewritten to be less restrictive. The format of the output
is less important than ensuring that the output contains the
intended URL, so the test now reflects that.