(RE-6473) Add accessors for structured platform data
This is in service of the effort to add granularity to remote
Apt repo operations -- we need to be able to define a list of
default Apt-using platforms (or releases) for a lot of our
remote Apt repo management jobs to use.

Right now, we shotgun all platforms based on hardcoded configs
on the remote infrastructure we run the commands on. But we have
a pretty comprehensive hash of basic platform data, and a lot
of adhoc accessors for getting data out of it.

This commit starts building some structured accessors we can
use to begin improving how we query or compare the data we're already
dragging along.