(RE-5817) Add more error handling
With FOSS_ONLY, we require FOSS_PLATFORMS to be set in the build-data
(or via environment variables), and the `pl:jenkins:retrieve` task needs
to be using the default remote target, otherwise we don't know how to
deal with the pathing. This will print out warnings in these cases and
prompt with a y/n for whether or not we should continue with retrieving
all of the available platforms.