(MAINT) Fix broken calls to ::File class methods
Namespacing is hard, and we use File#directory? and File#exist?
all over Packaging. When I tried fixing them, I gave up around
the 170th instance of File#somethingorother and instead made
some passthrough methods in the existing Pkg::Util::File class.

This is a dirty hack, and it shouldn't be how we do business.
But it'll unblock things while we rename the Pkg::Util classes
that conflict with core or stdlib classes.