(RE-6071) A small logic cleanup for config reassignment
There's now less inversion and less `#nil?`. This hopefully
provides a little bit more clarity for when this code path
is invoked. This commit also swaps `and` for `&&` to fix the
precedent order (there was a chance that without explicit parenthesis
around some of the logic previously that the compounding would be greedier
to the left than split across both conditions).