(RE-6071) Add rsync_cmd constructor
Generate arbitrary rsync commands for fun and profit!

The #rsync_cmd method should hopefully providea more abstract method
that we can eventually use to replace some of the hardcoded rsync commands
in other classes (like the Rpm and Deb classes). Some of the ugliness
in this method is due to being targeted at Ruby 1.9.3, so when we can
finally move to Ruby 2.1.x or newer this is one of the methods we should
revisit.