(RE-6071) Remove staging_server values for nuget and gems
Those two platforms are the least like any other package type we ship,
so the internal staging server pattern doesn't quite work for them. We
could probably squash these commits out of the history but I think
it's important to leave this commit in as a record because we need to
revisit how we ship (and therefore how we stage) these package formats.