(MAINT) Ensure #rsync_to conforms to new arguments format
`extra_flags` is kind-of a keyword argument now (thanks to the backwards
compatable, and annoying, implementation of an options hash in the
method definition). And by adding a call to ENV['DRYRUN'] in the #rsync_exec
method, we can ensure that #rsync_to and #rsync_from respect the dry-run
state without a million explicit calls to #rsync_to and #rsync_from having
"dryrun: ENV['DRYRUN']" appended to them.