(MAINT) Rub a little dry-run on it
Pkg::Util::Execution.ex doesn't respect ENV['DRYRUN']. Neither does
shipping Ruby Gems. This fixes that, and makes it a little easier
for us to start digging into the long work of making Packaing better
for everyone.

Some of these edits are not great but without gutting out all
of Pkg::Util::Execution and replacing it with Open3 (which is
what we should have done in 2014), this will get the job done.