(MAINT) Fix up tarball deployment
There is no Pkg::Config.tarball_path value, really. Tarballs are
discovered using a really gnarly file glob, so referring to that
value brings nothing but misery. For now, Tarballs may need to be
deployed with a light touch (manually) because that gnarly glob
will also pick up the packaging bundle that gets shipped as part
of the release process.