(RE-4366) Allow automatic MSI shipping
This allows MSI packages to be automatically shipped to the msi_host and
msi_staging_server locations, provided those variables have been set.

This commit ensures the only MSI packages that will actually be shipped
are those that are versioned appropriately. Packages without a version
or with the wrong version will be ignored. Please note, this glob does
not correctly match development packages so will not ship development
packages.