(maint) Only ship to jenkins-builds once
This commit removes the first invocation of the jenkins ship task to
builds.delivery.puppetlabs.net. This creates a shipped directory, and
places all the signed packages that were shipped to the public repos in
that directory. This should only happen once and it should only happen
after all the shipping and deploying has finished.