(maint) Add subtasks for pe_repo during platform removal
Adds two more subtasks to track deprecating pe_repo platform classes,
and to file a future ticket to remove them completely in the next PE
major.  This came up as a result of our forgetting to deal with pe_repo
during the removal of Debian6, Ubuntu 15.04 and 15.10, and the dance
around upgrades and installations which may have these platform classes
already declared.