RE-6989) Add component tag check
This adds a component tagging check for packaging. This (for now) is only a
human check and not a barrier for any CI jobs or builds using
pl:jenkins:uber_build. This commit adds the vanagon:check_tags and
vanagon:build_all rake tasks. The latter of which should replace the human
invocation of pl:jenkins:uber_build when we ship vanagon projects. The new
vanagon:build_all task is really just a wrapper for uber_build that adds
check_tags and a prompt.

Basically this is a human checklist, meant to notify the user if something
"looks tagged". It will not prevent shipment or builds, it will simply output
what "looks tagged" and then ask if this looks correct.