(maint) Add ticket to compile facter with java
I keep forgetting to ensure facter is compiled for new platforms with
java. This isn't an issue until we try to add a server package for the
new platform, and everything fails because the server package can't
utilize the facter available via puppet-agent. Rather than try to
remember this, let's just be explicit and get facter.jar for all
platforms.