(RE-6147) Fix recursive links in internal puppet-agent ship
When running the nightly_repos.rake file there would be a circular symlink and directories would be recursively made indefinitely. This fix takes out one of the calls to create a symlink (because it was called twice and should not have been). We also updated the two other existing symlink calls to be in the correct format. The symlink is set properly and the correct directories are made in the correct locations.