(maint) Move reprepro repo creation to aptly
This commit changes how we create one-off repos for development use.
Prior to this, we had been using reprepro to create and then later to
sign repos. These repos contain development builds on
builds.delivery.puppetlabs.net, nightlies.puppetlabs.com,
agent-downloads.delivery.puppetlabs.net, and s3.

Some of the repos, such as those on s3, do contain final builds made
available to our enterprise customers.

This commit switches out reprepro for aptly. Aptly handles repos
slightly differently, which adds a little complexity to this work. We
have to explicitly create the repo, add packages to it, and publish the
repo. In order to maintain the same directory structure that was
generated with reprepro, we must then link the dists and pool
directories into our working directory. This is done with a relative
link to enable these repos to be tar'd up and delivered to different
locations (i.e., s3). I do hope the repos are handled as a unit rather
than individually (i.e., the entire apt directory is moved rather than
the subdirectories contained within the apt directory).

These created repos are initially not signed. Signing happens later in
the process. This is done by updating the published repo with the needed
gpg key. This commit adds in a provision that checks to make sure the
repo we are signing is an aptly repo. If it isn't, it will default back
to the reprepro way of signing a repo. This should not be needed for
long, but it a holdover in case we do want to ship repos that had been
created prior to this change going in.