(PA-467) Add windows versionless symlinks to ship jobs
This is functionality to add the project-arch.msi symlinks that have no
version. The code itself looks for the symlink (if it's there already) and if
the source package exists before symlinking.

This should not break any legacy builds of puppet-agent as the lookup for the
existing file will prevent overwriting anything