(PA-467) More linking work
So, there's a lot of history here. We set packages to be immutable as
part of the shipping so they don't get overwritten, which doesn't play
well with symlinks. So we need to use FileUtils.ln(). Also,
File.symlink() and FileUtils.ln() behave differently with respect to
whether or not they expect the path for the source to be relative to pwd
or relative to the link target. By using hard links we should be able to
set them to be immutable without breaking everything.