(maint) Unlink only if it's actually a broken link
Previously, this could would attempt to unlink if the link didn't even
exist or if it was truly a broken symlink. This commit updates the check
to be sure there is actually a link that points to a non-existent file.
This means if we're creating a brand new set of links, then our
automation won't bomb out on us.