(RE-8179) Make ubership smarter
In the past we tried to make ubership smart enough to only update
metadata if packages existed. There were some issues with the initial
implementation, but the fix for that ended up undoing the initial work.

Dir.glob() always returns an array, and if it returns an empty array it
will never get into the unless statement. Convert this to check for
empty arrays.