(RE-5982) Add ship check task
This commit adds pl:ship_check to verify that shipping should work
before we actually try to ship. This will check things such as ssh
access to the necessary boxes, gpg keys available everywhere they are
needed, the needed gems are installed, and ownership is correct on the
gems on rubygems.