(PE-16465) Make a tarball of all signed repos
This adds a set of tasks to make a tarball of *all* of the
prepared repos, in addition to the individual per-platform tarballs.

This is useful for cutomers who have a PE master with
no internet access. They can unpack the puppet-agent-all tarball into the location that
pe_repo expects and use simplified agent install without needing internet access, or having to
manually download each agent that they need to feed to pe_repo.