(MAINT) pl:fetch is ☠️
The way we use a Git repo as a data service is a bad time, but we
don't need to make it worse by only being able to access it from a
Rake task. The BuildParams class encapsulates all of the logic that
was previously wrapped up in a Rake task, and is now required
alongside the Pkg::Config class whenever 'packaging' is required.

As a side effect, `pl:load_extras` no longer exists and is no longer
required to get the data into Pkg::Config when jobs depend on
`pl:fetch`.

This is probably not the complete refactoring that this code
deserves, but it should be a reasonable step-one in making our
configuration more deterministic and lowering the cognitive load
required to understand how Packaging's configuration works.