(QENG-4740) Call in to pkg directory in internal ship jobs
The 'generate_signed_repos' task did not previously execute out of the pkg
directory. This meant that in a later step, when 'prepare_signed_repos'
attempted to clean work that was no longer required, there was nothing to
clean.

This PR remedies that, so the job's disk footprint can be drastically
decreased.