(RE-5392) Ship ezbake.manifest if it exists
This commit also renames ezbake.manifest to include some unique build
information so we can tell which builds these files belong to.

This commit also looks for a yaml version of the ezbake manifest file,
names it with the build sha, and makes that file available as well.

This does not fail if either of those files do not exist.