(RE-8707) Add `platform_data`
This commit adds a `platform_data` array that maps each platform to where its
repo_config (if applicable) and artifact can be found. This info will get added
to the <ref>.yaml file that gets generated when a package is built, allowing
beaker to consume this info without having to figure out the mapping itself.