(RE-8836) Fix the final version check in packaging
This commit updated how the packaging repo checks if a version is final
or not. I'm removing quite a bit of obsolete code that's been hanging on
for much too long. We no longer use odd/even version schemes in any
project. It's also safe to combine some of these checks around which
versions are final into one place. We're effectively only parsing the
version string to ascertain that information.

We're also taking advantage of this major change to rename a few methods
and to update the code based on the current rspec, rubocop, and
rubocop-rspec suggestions. Though we are not making those updates here,
this will make our lives a little easier when we do update those checks.