(maint) Add a utility to fail if not git repo
This commit removes some duplicated code. Prior, we had been checking to
see if we were working in a git repo, and thus could use git commands.
If we weren't, we'd fail out with no real error message. Rather than
editing all instances of where we had implemented this check, this
commit adds in a utility to do that work. We can now fail uniformly with
this check.